+++ src/TortoiseProc/RevisionGraph/StandardLayoutNodeList.cpp
@@ -1,6 +1,6 @@
 // TortoiseSVN - a Windows shell extension for easy version control
 
-// Copyright (C) 2003-2008 - TortoiseSVN
+// Copyright (C) 2003-2009 - TortoiseSVN
 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -212,7 +212,7 @@ CString CStandardLayoutNodeList::GetToolTip (index_t index) const
             }
 
             tags +=   _T("\r\n")
-                    + CString (' ', tag->GetDepth() * 6) 
+                    + CString (' ', (int)tag->GetDepth() * 6) 
                     + tagInfo;
         }
