+++ src/TortoiseProc/RevisionGraph/StandardNodeSizeAssignment.cpp
@@ -63,7 +63,7 @@ void CStandardNodeSizeAssignment::ApplyTo (IRevisionGraphLayout* layout)
             size_t visibleElementCount = node->node->GetPath().GetDepth() 
                                        - node->skipStartPathElements
                                        - node->skipTailPathElements;
-            height += 3 + visibleElementCount * 21;
+            height += (int)(3 + visibleElementCount * 21);
         }
 
         // shift (root) nodes down, if their source has been folded
